split click commands from implementations
