+++ src/instructlab/cli/data/list.py
@@ -2,9 +2,8 @@
 
 # Standard
 from pathlib import Path
-from typing import List, TypedDict
+from typing import List
 import logging
-import time
 
 # Third Party
 import click
@@ -12,19 +11,10 @@ import click
 # First Party
 from instructlab import clickext
 from instructlab.configuration import DEFAULTS
-from instructlab.utils import convert_bytes_to_proper_mag, print_table
+from instructlab.data.list_data import list_data  # type: ignore
+from instructlab.utils import print_table
 
 logger = logging.getLogger(__name__)
-
-
-class DatasetListing(TypedDict):
-    """
-    Represents a single dataset listing.
-    """
-
-    filename: str
-    created_at: str
-    file_size: str
 
 
 @click.command(name="list")
@@ -36,42 +26,18 @@ class DatasetListing(TypedDict):
     default=[DEFAULTS.DATASETS_DIR],
     show_default=True,
 )
-@click.pass_context
-def list_data(ctx, dataset_dirs: list[str]):
+def list_datasets(dataset_dirs):
     """lists datasets"""
-    data: List[List[str]] = []
-    for directory in dataset_dirs:
-        dirpath = Path(directory)
-        directories: List[Path] = [dirpath]
-        top_level_dir = dirpath
-        if not dirpath.exists():
-            ctx.fail(f"directory does not exist: {dirpath}")
-        while len(directories) > 0:
-            current_dir = directories.pop()
-            for entry in current_dir.iterdir():
-                if entry.is_dir():
-                    directories.append(entry)
-                    continue
-                if entry.suffix != ".jsonl":
-                    continue
 
-                stat = entry.stat()
-                created_at = time.strftime(
-                    "%Y-%m-%d %H:%M:%S", time.localtime(stat.st_ctime)
-                )
-                fsize, mag = convert_bytes_to_proper_mag(stat.st_size)
-                formatted_size = f"{fsize:.2f} {mag}"
+    data: List[List[Path]] = []
 
-                # capture the entries as 'messages.jsonl', 'dir/dataset2.jsonl'
-                # with respect to the top-level directory
-                relative_name = str(entry.relative_to(top_level_dir))
-                data.append(
-                    [
-                        relative_name,
-                        created_at,
-                        formatted_size,
-                    ]
-                )
+    dirs = [Path(dir) for dir in dataset_dirs]
+
+    try:
+        data = list_data(dirs)
+    except OSError as exc:
+        click.secho(f"Failed to list datasets with exception: {exc}")
+        raise click.exceptions.Exit(1)
 
     headers = ["Dataset", "Created At", "File size"]
     print_table(headers, data)
